src/python/requirements.txt
oracledb==3.3.0
python-dotenv==1.1.1
python_bcrypt==0.3.2
